Refactorizar codigo en paquete padre
