+++ formatter.go
@@ -4,13 +4,9 @@ import (
 	"os"
 	"strings"
 	"time"
-)
 
-// formater doc ...
-type formater interface {
-	Development(buf *[]byte, t time.Time, file string, line int, function string, p Priority, s string, stack [][]byte)
-	Production(buf *[]byte, t time.Time, file string, line int, function string, p Priority, s string, stack [][]byte)
-}
+	"github.com/jgolang/log/core"
+)
 
 // These flags define which text to prefix to each log entry generated by the Logger.
 // Bits are or'ed together to control what's printed.
@@ -40,15 +36,15 @@ const (
 	LstdProdFlags  = Ldate | Ltime | Lmicroseconds | Lcaller | JSONFormat                                     // initial values for the standard production logger
 )
 
-// Formater doc ...
-type Formater struct {
+// Formatter doc ...
+type Formatter struct {
 	devFlag        int
 	prodFlag       int
 	additionalInfo string
 }
 
 // Production doc
-func (f Formater) Production(buf *[]byte, t time.Time, file string, line int, function string, p Priority, s string, stack [][]byte) {
+func (f Formatter) Production(buf *[]byte, t time.Time, file string, line int, function string, p core.Priority, s string, stack [][]byte) {
 	if f.prodFlag&(TerminalFormat) != 0 {
 		f.terminalFormat(buf, t, file, line, function, p, s, stack)
 		return
@@ -60,7 +56,7 @@ func (f Formater) Production(buf *[]byte, t time.Time, file string, line int, fu
 //   * l.prefix (if it's not blank),
 //   * date and/or time (if corresponding flags are provided),
 //   * file and line number (if corresponding flags are provided).
-func (f Formater) Development(buf *[]byte, t time.Time, file string, line int, function string, p Priority, s string, stack [][]byte) {
+func (f Formatter) Development(buf *[]byte, t time.Time, file string, line int, function string, p core.Priority, s string, stack [][]byte) {
 	if f.devFlag&(JSONFormat) != 0 {
 		f.jsonFormat(buf, t, file, line, function, p, s, stack)
 		return
@@ -68,26 +64,10 @@ func (f Formater) Development(buf *[]byte, t time.Time, file string, line int, f
 	f.terminalFormat(buf, t, file, line, function, p, s, stack)
 }
 
-func itoa(buf *[]byte, i int64, wid int) {
-	// Assemble decimal in reverse order.
-	var b [20]byte
-	bp := len(b) - 1
-	for i >= 10 || wid > 1 {
-		wid--
-		q := i / 10
-		b[bp] = byte('0' + i - q*10)
-		bp--
-		i = q
-	}
-	// i < 10
-	b[bp] = byte('0' + i)
-	*buf = append(*buf, b[bp:]...)
-}
-
-func (f Formater) jsonFormat(buf *[]byte, t time.Time, file string, line int, function string, p Priority, s string, stack [][]byte) {
+func (f Formatter) jsonFormat(buf *[]byte, t time.Time, file string, line int, function string, p core.Priority, s string, stack [][]byte) {
 	*buf = append(*buf, '{')
 	*buf = append(*buf, "\"level\":\""...)
-	*buf = append(*buf, GetTypeMsg(true, p)...)
+	*buf = append(*buf, core.GetTypeMsg(true, p)...)
 	*buf = append(*buf, "\",\"ts\":\""...)
 	ts := t.Unix()
 	itoa(buf, ts, 10)
@@ -145,7 +125,7 @@ func (f Formater) jsonFormat(buf *[]byte, t time.Time, file string, line int, fu
 	}
 }
 
-func (f Formater) terminalFormat(buf *[]byte, t time.Time, file string, line int, function string, p Priority, s string, stack [][]byte) {
+func (f Formatter) terminalFormat(buf *[]byte, t time.Time, file string, line int, function string, p core.Priority, s string, stack [][]byte) {
 	if f.devFlag&(Ldate|Ltime|Lmicroseconds) != 0 {
 		if f.devFlag&LUTC != 0 {
 			t = t.UTC()
@@ -174,7 +154,7 @@ func (f Formater) terminalFormat(buf *[]byte, t time.Time, file string, line int
 		*buf = append(*buf, '\t')
 	}
 	if f.devFlag&(Llevel) != 0 {
-		*buf = append(*buf, GetTypeMsg(false, p)...)
+		*buf = append(*buf, core.GetTypeMsg(false, p)...)
 		*buf = append(*buf, '\t')
 	}
 	if f.devFlag&(Lcaller) != 0 {
@@ -210,4 +190,68 @@ func (f Formater) terminalFormat(buf *[]byte, t time.Time, file string, line int
 		}
 		*buf = append(*buf, "\n---"...)
 	}
+}
+
+func itoa(buf *[]byte, i int64, wid int) {
+	// Assemble decimal in reverse order.
+	var b [20]byte
+	bp := len(b) - 1
+	for i >= 10 || wid > 1 {
+		wid--
+		q := i / 10
+		b[bp] = byte('0' + i - q*10)
+		bp--
+		i = q
+	}
+	// i < 10
+	b[bp] = byte('0' + i)
+	*buf = append(*buf, b[bp:]...)
+}
+
+func getLastStrSlice(s string, substr byte, jump int) string {
+	var str string
+	flag := 0
+	for i := len(s) - 1; i > 0; i-- {
+		if s[i] == substr {
+			if flag < jump {
+				flag++
+				continue
+			}
+			str = s[i+1:]
+			break
+		}
+	}
+	return str
+}
+
+func getLastToFirstStrSlice(s string, substr byte, jump int) string {
+	var str string
+	flag := 0
+	for i := len(s) - 1; i > 0; i-- {
+		if s[i] == substr {
+			if flag < jump {
+				flag++
+				continue
+			}
+			str = s[:i]
+			break
+		}
+	}
+	return str
+}
+
+func getFirstStrSlice(s string, substr byte, jump int) string {
+	var str string
+	flag := 0
+	for i := 0; i < len(s); i++ {
+		if s[i] == substr {
+			if flag < jump {
+				flag++
+				continue
+			}
+			str = s[:i]
+			break
+		}
+	}
+	return str
 }
