+++ go.mod
@@ -1,3 +1,5 @@
 module github.com/jgolang/log
 
 go 1.15
+
+require github.com/jgolang/log/core v0.0.0-20210112060917-fc4af39e5bd8
